# perf(ext/web): avoid reallocations in op_base64_atob