Follow-up to #15008 

op_base64_atob did SmallVec => Vec and vice-versa. 
This commit makes it use SmallVec directly.

`main` -
```
b64Short:       n = 1000000,            dt = 0.225s,    freq = 4438775.950/s,   time = 225ns/op
```

this patch -
```
b64Short:       n = 1000000,            dt = 0.185s,    freq = 5404132.279/s,   time = 185ns/op
```